requirements.txt
gymnasium>=0.29.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
Pillow>=9.0
